Fixed test cases. Small amount of refactoring.
