+++ src/can/Can.ts
@@ -2,6 +2,7 @@ import { InvalidPayloadLength } from './Errors';
 import BitUtils from './BitUtils';
 import { EndianType } from '../shared/DataTypes';
 import { DbcData, Message, Signal } from '../dbc/DbcTypes';
+import { BoundMessage, BoundSignal, Frame, Payload } from './CanTypes';
 
 /**
  * The Can class offers utility functions that aid in the processing of general CAN data/information
@@ -55,8 +56,10 @@ class Can extends BitUtils {
     // Uint8Array will ensure values don't exceed 255
     const byteArray = new Uint8ClampedArray(payload);
 
+    // Set extended flag
+    const canId = isExtended ? this.setExtendedFlag(id) : id;
     return {
-      id,
+      id: canId,
       dlc: payload.length,
       isExtended,
       payload: Array.from(byteArray),
@@ -294,34 +297,3 @@ class Can extends BitUtils {
   }
 }
 export default Can;
-
-export type Frame = {
-  id: number;
-  dlc: number;
-  isExtended: boolean;
-  payload: Payload;
-};
-
-export type BoundMessage = {
-  boundData: {
-    message: Message;
-    frame: Frame;
-  };
-  boundSignals: Map<string, BoundSignal>;
-  id: number;
-  name: string;
-  setSignalValue: (signal: string, value: number) => {};
-};
-
-export type BoundSignal = {
-  boundData: {
-    signal: Signal;
-    payload: Payload;
-  };
-  value: number;
-  rawValue: number;
-  physValue: string;
-  setValue: (value: number) => {};
-};
-
-export type Payload = number[];
